refactor: cleaning structure in dev
