+++ dev/db/seed12.ts
@@ -5,7 +5,7 @@ import { readFileSync } from 'node:fs';
 import postgres from 'postgres';
 
 import { characters, relations } from '../../src/lib/server/db/schema/schema.ts';
-import { assertNumberEquals } from '../assertions.ts';
+import { assertNumberEquals } from '../utils/assertions.ts';
 
 const env = parse(readFileSync('.env'));
 if (!('DATABASE_URL' in env)) throw new Error('DATABASE_URL not found in .env!');
