+++ package.json
@@ -10,7 +10,7 @@
     "db:sync": "npm run db:generate && npm run db:migrate",
     "db:studio": "drizzle-kit studio",
     "firstrun:dev": "npm install && npm run check && docker compose up --detach --wait && npm run db:sync && npm run seed:dev",
-    "seed:dev": "node ./dev/db/seed10.ts",
+    "seed:dev": "node ./dev/db/seed12.ts",
     "preview": "vite preview",
     "autofix": "npm run lint:fix && npm run format",
     "commit": "cz",
